+++ spec/lib/fixtures/source_location_test.rb
@@ -1,0 +1,7 @@
+
+
+
+# Do not move this method
+# This file is used to test Method#source_location
+def m
+end