+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Whitespace conventions:
     * `Array#sample`
 - Added rudimental history support to `opal-repl`, just create the history file (`~/.opal-repl-history`) and it record the last 1000 lines
 - Added `JS::Error` error class that can be used to catch any JS error.
+- Added `Method#source_location` and `Method#comments`.
 
 
 ### Changed